fix: container moved outside of the App.jsx
